github: extract into separate library
Preparation for making it simpler to add per_page=100 to URLs. And for
unifying the auth method (github.Client.Get vs github.DoAuthRequest).

This basically served mostly for me to be able to understand my own code
again.

NOTE: I did not test all commands simply because I don't have tests
(integration tests like google/github-go would be really nice). I tried a
couple of things when I first extracted the library and they were broken. So
this may yet break people. I won't tag a release yet, and daredevils who do
go get will hopefully report bugs).

Updates #45.
