+++ github-release.go
@@ -4,8 +4,11 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/aktau/github-release/github"
 	"github.com/voxelbrain/goptions"
 )
+
+const GH_URL = "https://github.com"
 
 type Options struct {
 	Help      goptions.Help `goptions:"-h, --help, description='Show this help'"`
@@ -111,6 +114,7 @@ func main() {
 	}
 
 	VERBOSITY = len(options.Verbosity)
+	github.VERBOSITY = VERBOSITY
 
 	if cmd, found := commands[options.Verbs]; found {
 		err := cmd(options)
