+++ releases.go
@@ -5,12 +5,13 @@ import (
 	"strings"
 	"time"
 
+	"github.com/aktau/github-release/github"
 	"github.com/dustin/go-humanize"
 )
 
 const (
-	RELEASE_LIST_URI    = "/repos/%s/%s/releases%s"
-	RELEASE_LATEST_URI  = "/repos/%s/%s/releases/latest%s"
+	RELEASE_LIST_URI    = "/repos/%s/%s/releases"
+	RELEASE_LATEST_URI  = "/repos/%s/%s/releases/latest"
 	RELEASE_DATE_FORMAT = "02/01/2006 at 15:04"
 )
 
@@ -67,11 +68,8 @@ type ReleaseCreate struct {
 }
 
 func Releases(user, repo, token string) ([]Release, error) {
-	if token != "" {
-		token = "?access_token=" + token
-	}
 	var releases []Release
-	err := GithubGet(fmt.Sprintf(RELEASE_LIST_URI, user, repo, token), &releases)
+	err := github.Client{Token: token, BaseURL: EnvApiEndpoint}.Get(fmt.Sprintf(RELEASE_LIST_URI, user, repo), &releases)
 	if err != nil {
 		return nil, err
 	}
@@ -80,11 +78,8 @@ func Releases(user, repo, token string) ([]Release, error) {
 }
 
 func latestReleaseApi(user, repo, token string) (*Release, error) {
-	if token != "" {
-		token = "?access_token=" + token
-	}
 	var release Release
-	return &release, GithubGet(fmt.Sprintf(RELEASE_LATEST_URI, user, repo, token), &release)
+	return &release, github.Client{Token: token, BaseURL: EnvApiEndpoint}.Get(fmt.Sprintf(RELEASE_LATEST_URI, user, repo), &release)
 }
 
 func LatestRelease(user, repo, token string) (*Release, error) {
