+++ tags.go
@@ -2,10 +2,12 @@ package main
 
 import (
 	"fmt"
+
+	"github.com/aktau/github-release/github"
 )
 
 const (
-	TAGS_URI = "/repos/%s/%s/tags%s"
+	TAGS_URI = "/repos/%s/%s/tags"
 )
 
 type Tag struct {
@@ -19,18 +21,8 @@ func (t *Tag) String() string {
 	return t.Name + " (commit: " + t.Commit.Url + ")"
 }
 
-/* get the tags associated with a repo */
+// Get the tags associated with a repo.
 func Tags(user, repo, token string) ([]Tag, error) {
 	var tags []Tag
-
-	if token != "" {
-		token = "?access_token=" + token
-	}
-
-	err := GithubGet(fmt.Sprintf(TAGS_URI, user, repo, token), &tags)
-	if err != nil {
-		return nil, err
-	}
-
-	return tags, nil
+	return tags, github.Client{Token: token, BaseURL: EnvApiEndpoint}.Get(fmt.Sprintf(TAGS_URI, user, repo), &tags)
 }
